pyproject.toml
[tool.poetry]
name = "famistudio_converter"
version = "0.1.0"
description = "Converts famistudio files into Python readable data"
authors = ["TheJoeSmo <joesmo.joesmo12@gmail.com>"]
readme = "README.md"
license = "GPL3"
include = [
    "LICENSE"
]

[tool.poetry.dependencies]
python = "^3.10"
pretty-errors = "^1.2.25"
attrs = "^22.1.0"

[tool.poetry.dev-dependencies]
isort = "^5.10.1"
pyupgrade = "^2.37.3"
black = "^22.6.0"
pytest = "^7.1.2"
flake8 = "^5.0.4"

[tool.isort]
profile = "black"
known_first_party = "foundry"

[tool.black]
line-length = 120

[tool.pytest]
qt_api = "PySide6"
ignore = "tests/game/gfx/objects/"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
